Fix silent float32 type promotion in mask_cond

+++ SwiftSketch/model/SwiftSketch_model.py
@@ -87,7 +87,7 @@ class SwiftSketch(nn.Module):
         if force_mask:
             return torch.zeros_like(cond)
         elif self.training and self.cond_mask_prob > 0.: 
-            mask = torch.bernoulli(torch.ones(bs, device=cond.device) * self.cond_mask_prob).view(bs, 1,1,1)  # 1-> use null_cond, 0-> use real cond
+            mask = torch.bernoulli(torch.ones(bs, device=cond.device) * self.cond_mask_prob).to(cond.dtype).view(bs, 1,1,1)  # 1-> use null_cond, 0-> use real cond
             return cond * (1. - mask)
         else:
             return cond
